feat: add car-view in winner record
