+++ src/components/winner/config.ts
