+++ src/components/winner/index.ts
@@ -1,1 +1,1 @@
-export * from './winner-view';
+export * from './view';
