+++ src/components/winner/style.module.scss
@@ -1,0 +1,45 @@
+.row {
+  border: 1px solid #0008;
+}
+
+.cell {
+  padding: 0 20px;
+  text-align: center;
+  height: 60px;
+  border-radius: 30px;
+  box-shadow: inset 0 0 30px 0 #fff1, inset 0 0 0 1px #fff1;
+  backdrop-filter: blur(5px);
+  color: #888;
+  font-weight: 400;
+  font-size: 35px;
+  font-family: var(--ff-mono);
+
+  &.num {
+    --reserved: 1;
+  }
+  &.car {
+    position: relative;
+    width: 140px;
+  }
+  &.name {
+    font-family: var(--ff-sans);
+    font-size: 20px;
+  }
+  &.wins {
+    --reserved: 1;
+  }
+  &.time {
+    --reserved: 1;
+  }
+}
+
+.small-bug {
+  --position: calc(50% - 40px) !important;
+  width: 80px;
+  height: 40px;
+}
+.hidden {
+  .small-bug {
+    opacity: 0;
+  }
+}
